backend/requirements.txt
# Django Core
Django==5.1.14
python-decouple==3.8
setuptools>=65.0.0

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-filter==24.3
django-cors-headers==4.6.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Caching & Tasks
redis==5.1.1
django-redis==5.4.0
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Authentication
djoser==2.3.0

# Security
django-ratelimit==4.1.0

# File Handling
Pillow==11.0.0

# Payments
stripe==11.1.1

# Monitoring (Optional)
sentry-sdk==2.19.2

# Development
django-debug-toolbar==5.1.0
faker==33.1.0

# Timezone
pytz==2024.2


django-extensions==3.2.3
setuptools>=70.0.0

gunicorn==21.2.0